requirements.txt
pkg-resources==0.0.0
github3.py==1.2.0